Fixed all failing tests after moving mention related stuff to repository
