+++ lib/cubits/comment/comment_cubit.dart
@@ -45,7 +45,7 @@ class CommentCubit extends Cubit<CommentState> {
           emit(CommentsLoaded(_comments));
         }
         final mentions = _comments
-            .map((comment) => getUserIdsInComment(comment.text))
+            .map((comment) => _repository.getUserIdsInComment(comment.text))
             .expand((mention) => mention)
             .toList();
         if (mentions.isEmpty) {
@@ -56,7 +56,8 @@ class CommentCubit extends Cubit<CommentState> {
             profilesList.map((profile) => MapEntry<String, Profile>(profile!.id, profile)));
         _comments = _comments
             .map((comment) => comment.copyWith(
-                text: replaceMentionsWithUserNames(profiles: profiles, comment: comment.text)))
+                text: _repository.replaceMentionsWithUserNames(
+                    profiles: profiles, comment: comment.text)))
             .toList();
         emit(CommentsLoaded(_comments));
       });
@@ -79,24 +80,12 @@ class CommentCubit extends Cubit<CommentState> {
       _comments.insert(0, comment);
       emit(CommentsLoaded(_comments));
       final mentions = _repository.getMentionedProfiles(comment.text);
-      final mentionReplacedText = replaceMentionsInAComment(comment: text, mentions: mentions);
+      final mentionReplacedText =
+          _repository.replaceMentionsInAComment(comment: text, mentions: mentions);
       await _repository.comment(text: mentionReplacedText, videoId: _videoId, mentions: mentions);
     } catch (err) {
       emit(CommentError(message: 'Error commenting.'));
     }
-  }
-
-  /// Called everytime comment is being edited
-  /// Checks if there are any mentions in a comment and returns suggestion
-  Future<void> getMentionSuggestion(String comment) async {
-    final mentionedUserName = getMentionedUserName(comment);
-    if (mentionedUserName == null) {
-      emit(CommentsLoaded(_comments));
-      return;
-    }
-    emit(CommentsLoaded(_comments, isLoadingMentions: true));
-    final mentionSuggestions = await _repository.getMentions(mentionedUserName);
-    emit(CommentsLoaded(_comments, mentionSuggestions: mentionSuggestions));
   }
 
   /// Called when mention suggestion has been tapped
@@ -112,46 +101,17 @@ class CommentCubit extends Cubit<CommentState> {
     return '${commentText.substring(0, lastSpaceIndex)} @$profileName ';
   }
 
-  /// Replaces mentioned user names with users' id in comment text
+  /// Called everytime comment is being edited
+  /// Checks if there are any mentions in a comment and returns suggestion
   @visibleForTesting
-  String replaceMentionsInAComment({required String comment, required List<Profile> mentions}) {
-    var mentionReplacedText = comment;
-    for (final mention in mentions) {
-      mentionReplacedText = mentionReplacedText.replaceAll('@${mention.name}', '@${mention.id}');
+  Future<void> getMentionSuggestion(String comment) async {
+    final mentionedUserName = _repository.getMentionedUserName(comment);
+    if (mentionedUserName == null) {
+      emit(CommentsLoaded(_comments));
+      return;
     }
-    return mentionReplacedText;
-  }
-
-  /// Extracts the username to be searched within the database
-  @visibleForTesting
-  String? getMentionedUserName(String comment) {
-    final mention = comment.split(' ').last;
-    if (mention.isEmpty || mention[0] != '@') {
-      return null;
-    }
-    final mentionedUserName = mention.substring(1);
-    if (mentionedUserName.isEmpty) {
-      return null;
-    }
-    return mentionedUserName;
-  }
-
-  @visibleForTesting
-  List<String> getUserIdsInComment(String comment) {
-    final regExp = RegExp(r'@[0-9a-f]{8}\b-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-\b[0-9a-f]{12}\b');
-    final matches = regExp.allMatches(comment);
-    return matches.map((match) => match.group(0)!.substring(1)).toList();
-  }
-
-  @visibleForTesting
-  String replaceMentionsWithUserNames({
-    required Map<String, Profile> profiles,
-    required String comment,
-  }) {
-    final regExp = RegExp(r'@[0-9a-f]{8}\b-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-\b[0-9a-f]{12}\b');
-    return comment.replaceAllMapped(
-        regExp,
-        (match) =>
-            '@${profiles[match.group(0)!.substring(1)]?.name ?? match.group(0)!.substring(1)}');
+    emit(CommentsLoaded(_comments, isLoadingMentions: true));
+    final mentionSuggestions = await _repository.getMentions(mentionedUserName);
+    emit(CommentsLoaded(_comments, mentionSuggestions: mentionSuggestions));
   }
 }
