+++ test/pages/tabs/notifications_tab_test.dart
@@ -21,7 +21,8 @@ void main() {
     testWidgets('Renders NotificationsTab correctly', (tester) async {
       final repository = MockRepository();
 
-      when(repository.getNotifications).thenAnswer((invocation) => Future.value([
+      when(repository.getNotifications).thenAnswer((invocation) => Future.value());
+      when(() => repository.notificationsStream).thenAnswer((_) => Stream.value([
             AppNotification(
               type: NotificationType.like,
               createdAt: DateTime.now(),
@@ -82,7 +83,8 @@ void main() {
     testWidgets('Takes care of new notification properly', (tester) async {
       final repository = MockRepository();
 
-      when(repository.getNotifications).thenAnswer((invocation) => Future.value([
+      when(repository.getNotifications).thenAnswer((invocation) => Future.value());
+      when(() => repository.notificationsStream).thenAnswer((invocation) => Stream.value([
             AppNotification(
               type: NotificationType.like,
               createdAt: DateTime.now(),
@@ -129,7 +131,8 @@ void main() {
     testWidgets('Empty notification will show empty text', (tester) async {
       final repository = MockRepository();
 
-      when(repository.getNotifications).thenAnswer((invocation) => Future.value([]));
+      when(repository.getNotifications).thenAnswer((invocation) => Future.value());
+      when(() => repository.notificationsStream).thenAnswer((invocation) => Stream.value([]));
 
       await tester.pumpApp(
         widget: BlocProvider<NotificationCubit>(
